pci: extend MSI interrupt vector number to 32 bits
Also increase MSI message data size to 32 bits according to PCIe spec.

Remove 0xff check for MSI interrupts because it is potentially valid
interrupt vector number. Reject 0xff only for legacy pin interrupts.

- MSI-X supports up to 2048 interrupts per device that do not fit to
`uint8`.

- Non-x86 systems may use separate interrupt vector ranges for
hard-wired interrupts and MSI interrupts so `uint8` is not enough to
represent all of them.

+++ headers/os/drivers/dm2/bus/PCI.h
@@ -35,15 +35,15 @@ public:
 	virtual void 		SetPowerstate(uint8 state) = 0;
 
 	// MSI/MSI-X
-	virtual uint8		GetMsiCount() = 0;
-	virtual status_t	ConfigureMsi(uint8 count, uint8* startVector) = 0;
+	virtual uint32		GetMsiCount() = 0;
+	virtual status_t	ConfigureMsi(uint32 count, uint32* startVector) = 0;
 	virtual status_t	UnconfigureMsi() = 0;
 
 	virtual status_t	EnableMsi() = 0;
 	virtual status_t	DisableMsi() = 0;
 
-	virtual uint8		GetMsixCount() = 0;
-	virtual status_t	ConfigureMsix(uint8 count, uint8* startVector) = 0;
+	virtual uint32		GetMsixCount() = 0;
+	virtual status_t	ConfigureMsix(uint32 count, uint32* startVector) = 0;
 	virtual status_t	EnableMsix() = 0;
 
 protected:
